XXX manager: when signalled to shutdown etc, use JOB_REPLACE_IRREVERSIBLY
This is required to ensure that shutdown is not interrupted by the
activation of a unit with a conflict.

We already used the correct job mode for `ctrl-alt-del.target`.  But not
for `exit.target` (SIGINT of user manager).  And not for `sigpwr.target` or
`kbrequest.target`, which are also candidates for linking to a shutdown
target.

XXX It's also possible the target is not a shutdown target, e.g. `sigpwr`
was linked to start a script which reads the old /etc/powerstatus logic.
Then switching to REPLACE_IRREVERSABLY would cause the script to fail
when running `systemctl poweroff`, if they had not known to set
DefaultDependencies=no.

SIGRTMIN+0 is also changed to isolate default.target, instead of starting
it.  The previous behaviour was documented, but we have no justification
to give for it.

XXX docs need updating