logind: change the logic used by verify_shutdown_creds() for simplicity
It's not clear why it would ignore inhibitors or multiple sessions if
PolKit actions are not passed for these - surely any "shutdown" type action
would require authorization to override these.  It would make more sense to
refuse, if we didn't know how to authorize it.  But we always do.

Change these NULL checks to assertions.

Looking a bit further, there is one weird case.  If you set a scheduled
shutdown, there's a gaping TOCTOU question.  It will be simpler to reason
about this, if we require pre-authorization to shut down in the face of
inhibitors and multiple sessions.  I believe this change is reasonable.
Scheduled `shutdown` is a historic command-line tool for multi-user
systems.  Users of it should need permission to bring the system down even
when there are other users logged in.  The old behaviour seems to fragile
to depend on _deliberately_, at least.  Also `shutdown` probably makes most
sense for SSH servers e.g. in universities, where it's _likely_ the admin
also logs in over SSH, and they wouldn't be granted permission though the
PolKit policy anyway.

It is also clearer to check all the relevant authorizations, instead of
treating `foo` as implicitly authorized when one or both of
`foo-multiple-sessions` or `foo-ignore-inhibit` applies.  Specifically,
it makes it easier to see that if both multiple-sessions and force-inhibit
would apply, then both must be authorized.