systemctl: scheduled shutdown does not fall back to immediate shutdown
Fix error message.  Users are expected to fall back to `shutdown now`.

I guess we are not backwards compatible if you disable logind, unlike the
old `systemd-shutdownd` implementation.  I'm not concerned about this
myself... though it's not clear whether there was a rationale for changing
it (e.g. people wanting to manage it under PolKit?).