[WIP] systemctl: can we clarify action_table?
action_table is not a table of all actions.

It's a mix of the actions from legacy commands, and verbs from systemctl,
which work by starting one of the special target units.  There's some
overlap as well.

I wondered if I could clarify it by separating the two.  That's not very
straightforward, so I ended up with something else.  I think it's OK but
am not clear it's worth the churn.