execute: allow RootImage= to lack some protected directories
Treat RootImage= the same as RootDirectory=, in allowing it to lack some
of the directories mentioned in ProtectSystem (e.g. /home).

I can't see the reason for treating RootImage= differently, so it just
jumps out as looking wrong.

Don't worry that I'm introducing more lines of code, I needed the extra
flag variable anyway for the next commit :).