execute: treat RootDirectory= (and RootImage=) failure as fatal
I think it's already such a natural assumption that the existing docs are
implying it, and not doing this was simply a bug.

So much that I gave up trying to document this more explicitly.  At least
the error message will point towards some problem related to privileges,
and to mount namespaces.  RootDirectory= is not used as widely as the other
namespace options, and running the system manager without the required
CAP_SYS_ADMIN is recommended against, so this isn't a massively prominent
issue.

That said, this is still slightly ugly.

The failure of setup_namespace() might be due specifically to
PrivateDevices= and running without CAP_MKNOD.  Then removing
PrivateDevices= could allow the unit to successfully launch inside the
chroot.  This is not consistent with the documented behaviour of
PrivateDevices= and related options.

This is a more noticeable extension of an existing eyesore.  If
PrivateDevices= fails, it will also skip some e.g. ProtectSystem=, though
not others e.g. PrivateTmp=, which is not documented.  Oh, unless you
specified RootDirectory= / RootImage=, in which case it will apply to all
mount namespace manipulation e.g. including PrivateTmp=.

Before this commit, it would have attempted to continue and launch
_outside_ the chroot.  Which might succeed, or succeed but break the admins
security policy when systemd did not explicitly document that this was a
possibility, or succeed in launching something _different_ which happens to
have the same name outside the chroot.