log: do not increase socket buffers for logging by default
This effectively reverts almost all of bb99a35a,
"log: increase socket buffers for logging by default".

If you're familiar with the bufferbloat project, seeing these will make you
suspicious.  And I think we only have a justification for one of them.

This is the only motive I have here.  Right now, there's one place where
the buffer size is arguably justified, and several others where it's not.
I want to remove the unneeded logic, to make it easier to analyze what we
really rely on.

Firstly, we're setting the receive buffer size on unix(7) sockets.  This
is explicitly documented to do nothing on Linux.  So we can remove it
and everything will work as before.

Increasing the buffer size used by stdout logging, or even basic/log.c,
does not seem very useful.  The previous commit assumes people will not
decrease the system default socket buffer size below 8kB.  And syslog() is
possibly the most widespread use of unix sockets; I think we can trust that
people who tweak the default buffer will be aware of syslog().

In the unlikely case of a deadlock, PID 1 will timeout log messages at a
rate up to 100 per second.  Other users of basic/log.c will hang for 10
seconds before skipping log messages when journald is not keeping up.  IMO
this works equally with either buffer size.  (And arguably better with the
tiny one, but I don't feel strongly enough to enforce that).

Finally, the native journald socket allows for larger messages.  AFAICT
this is to support serializing core dumps in to the log stream.  To
support this, the native journald API detects when the datagram was too big
for the socket buffer, and sends a memfd instead, defeating the normal
buffering control :).  This doesn't require an additional data copy if
there is some message that is larger than the buffer would permit, though
it does involve extra system calls.

I've kept the larger buffer for forwarding to syslog.socket.  This one is
arguably useful.  If syslog.service doesn't keep up, and the socket buffer
becomes full, systemd-journald will skip forwarding it messages until there
is space again.  In particular journald starts forwarding very early on,
but syslog.service likely does not start until after sysinit.target.

I've kept the `ReceiveBuffer=` setting in syslog.socket as a more visible
reminder, until we warn and document that this has no effect on unix
sockets (see #8158).