manager: do not increase buffer for cgroups-agent socket
It's impossible for data to be buffered in the receiving socket on Linux.
It's explicitly documented that the receive socket buffer size SO_RCVBUF
has no effect on unix(7) sockets.  Code that is portable to Linux must
allow for this.  And much the same for the BSD implementation.
https://patchwork.freedesktop.org/patch/23219/

Happily, we had no reason for increasing the buffer size here :).

cgroups-agent only sends a small datagram to name the cgroup which became
empty. It already assumes the datagram is small enough to be permitted by
the default send buffer size.  We could have lots of cgroups become empty
at once, however cgroups-agent is run by the kernel (lol) for each
individual one.  Hence each datagram will be queued in different sending
socket buffer.

I guess it might be nice if you could see the cgroup-agent processes pile
up... but that would imply setting SO_LINGER or something.  It seems to be
getting very obscure, so I will leave the code as simple as possible.