units: Shut down the journal cleanly before switch-root
switch-root kills all outstanding processes.  Immediately before that,
let's cleanly stop the journal.  I don't think SIGKILL can be good for
those binary file formats :) and for any log messages in the process of
being written to the filesystem.

So neither of these steps is going to be logged to the journal :).
My answer is to disable logging for both these steps, but make them
separate services so that PID 1 will log their status (falling back to
kmsg).

I have allowed the boot to continue if initrd-shutdown-journald.service
fails.  The implementation could be much simpler if I didn't do this...
and maybe easier to maintain... but I am an anxious person :).  I think
a Wants= dependency is valid here because the boot does not strictly
require the integrity of the journal.  And our boots seem to have
worked without this as a guarantee so far :).