timedate: fix message text for systemd-timesyncd.service status
timedatectl can show "systemd-timesyncd.service active: yes", but "active"
has a specific meaning in systemd.  What timedated actually only tests
whether the service is _enabled_.  It may have been enabled and yet
stopped (or failed to start).  It may also have been started but not
enabled, in which case you could be very confused why systemd-timesyncd
is not "active", but when you manually set the date something is still
immediately resetting it from network time.

(It turns out that part is not why I was confused.  It was a can of worms
with timedatex, particularly after the recent change to this message. It
has been pointed out to Fedora now and I'll see what happens
https://bugzilla.redhat.com/show_bug.cgi?id=1550259
Of course if anyone thinks we _would_ accept some fix upstream, I'd be very
happy as a Fedora user and do not want to stand in the way.  I didn't have
time to work out anything more solid now.

So here's this hunk of code; if it does get totally super-ceded by wider
changes that's also fine and I'll probably be even happier).